% file gllm/lsat.Rd
%
\name{lsat}
\alias{lsat}
\title{
Five dichotomous items from the Law School Admission Test (LSAT)
}
\description{
Small dataset (items 11-15 from sections 6 and 7 of the Law School Admission
Test) used by Bock and Lieberman (1970), Christoffersson (1975)
and Joreskog and Sorbom (1986) to test methods for factor analysis
of binary data.
}

\source{
Bock RD, Lieberman M (1970).  Fitting a response model for n dichotomously
scored items.
\emph{Psychometrika}
\bold{35}, 179-197.
}
\keyword{datasets}
