\name{print.summary.gllm}
\alias{print.summary.gllm}
\title{Print results of gllm}
\description{
Print contents of result of call to \code{gllm}.
}
\usage{
print.summary.gllm(x,digits=NULL,show.residuals=FALSE,...) 
}
\arguments{
  \item{x}{is the object output from gllm.}
  \item{digits}{is the number of output digits.}
  \item{show.residuals}{toggles printing of the fitted values and residuals.}
  \item{...}{Additional arguments to be passed to \code{print.generic}.}
}
\keyword{ print
}
